Added dependency to reveal-highlight-themes.
This will pull in the available Highlight.js themes as a Bower
dependency and make them available for usage in the presentation.

The available syntax themes can be found in the
`bower_components/reveal-highlight-themes/styles` folder and can be
selected by changing the stylesheet reference in the `_index.html` file
in the generated presentation.