Document how to install dependencies for existing presentations
I wanted to update an older presentation for an upcoming conference but was confused for a while because `grunt serve` was showing me an empty page without throwing any errors. I missed the `bower install` step.

Maybe we can provide some information in the README on how to work on existing presentations?